refactor: move ring implementations to dedicated module
