+++ src/crypto/key_expansion.rs
@@ -7,64 +7,13 @@ use crate::error::Result;
 #[derive(Debug, Default, Clone, Copy)]
 pub struct KeyMaterial<'a>(pub &'a [u8]);
 
-impl KeyMaterial<'_> {
-    pub fn expand_as_secret(&self, cipher_suite: &CipherSuite) -> Result<Secret> {
-        ring::expand_secret_from(self.0, cipher_suite)
-    }
+pub trait ExpandAsSecret {
+    fn expand_as_secret(&self, cipher_suite: &CipherSuite) -> Result<Secret>;
 }
 
-const SFRAME_HKDF_SALT: &[u8] = "SFrame10".as_bytes();
-const SFRAME_HKDF_KEY_EXPAND_INFO: &[u8] = "key".as_bytes();
-const SFRAME_HDKF_SALT_EXPAND_INFO: &[u8] = "salt".as_bytes();
-
-mod ring {
-    use crate::{
-        crypto::{
-            cipher_suite::{CipherSuite, CipherSuiteVariant},
-            secret::Secret,
-        },
-        error::{Result, SframeError},
-    };
-
-    use super::{SFRAME_HDKF_SALT_EXPAND_INFO, SFRAME_HKDF_KEY_EXPAND_INFO, SFRAME_HKDF_SALT};
-
-    struct OkmKeyLength(usize);
-
-    impl ring::hkdf::KeyType for OkmKeyLength {
-        fn len(&self) -> usize {
-            self.0
-        }
-    }
-
-    impl From<&CipherSuite> for ring::hkdf::Algorithm {
-        fn from(cipher_suite: &CipherSuite) -> Self {
-            match cipher_suite.variant {
-                CipherSuiteVariant::AesGcm128Sha256 => ring::hkdf::HKDF_SHA256,
-                CipherSuiteVariant::AesGcm256Sha512 => ring::hkdf::HKDF_SHA512,
-            }
-        }
-    }
-
-    pub fn expand_secret_from(key_material: &[u8], cipher_suite: &CipherSuite) -> Result<Secret> {
-        let algorithm = cipher_suite.into();
-        let prk = ring::hkdf::Salt::new(algorithm, SFRAME_HKDF_SALT).extract(key_material);
-
-        let key = expand_key(&prk, SFRAME_HKDF_KEY_EXPAND_INFO, cipher_suite.key_len)?;
-        let salt = expand_key(&prk, SFRAME_HDKF_SALT_EXPAND_INFO, cipher_suite.nonce_len)?;
-
-        Ok(Secret { key, salt })
-    }
-
-    fn expand_key(prk: &ring::hkdf::Prk, info: &[u8], key_len: usize) -> Result<Vec<u8>> {
-        let mut sframe_key = vec![0_u8; key_len];
-
-        prk.expand(&[info], OkmKeyLength(key_len))
-            .and_then(|okm| okm.fill(sframe_key.as_mut_slice()))
-            .map_err(|_| SframeError::KeyExpansion)?;
-
-        Ok(sframe_key)
-    }
-}
+pub const SFRAME_HKDF_SALT: &[u8] = "SFrame10".as_bytes();
+pub const SFRAME_HKDF_KEY_EXPAND_INFO: &[u8] = "key".as_bytes();
+pub const SFRAME_HDKF_SALT_EXPAND_INFO: &[u8] = "salt".as_bytes();
 
 #[cfg(test)]
 mod test {
@@ -77,6 +26,8 @@ mod test {
         },
         util::test::assert_bytes_eq,
     };
+
+    use super::ExpandAsSecret;
 
     fn derive_correct_keys(variant: CipherSuiteVariant) {
         let test_vector = get_test_vector(&variant.to_string());
