+++ src/crypto/ring/mod.rs
@@ -2,8 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0 AND MIT
 
 pub mod aead;
-pub mod cipher_suite;
 pub mod key_expansion;
-pub mod secret;
-
-mod ring;
