+++ src/receiver.rs
@@ -7,7 +7,7 @@ use crate::{
     crypto::{
         aead::AeadDecrypt,
         cipher_suite::{CipherSuite, CipherSuiteVariant},
-        key_expansion::KeyMaterial,
+        key_expansion::{ExpandAsSecret, KeyMaterial},
         secret::Secret,
     },
     error::{Result, SframeError},
